+++ packages/atomic/src/components/commerce/atomic-commerce-interface/e2e/fixture.ts
@@ -1,0 +1,18 @@
+import {test as base} from '@playwright/test';
+import {
+  AxeFixture,
+  makeAxeBuilder,
+} from '../../../../../playwright-utils/base-fixture';
+import {CommerceInterfacePageObject as CommerceInterface} from './page-object';
+
+type MyFixtures = {
+  commerceInterface: CommerceInterface;
+};
+
+export const test = base.extend<MyFixtures & AxeFixture>({
+  makeAxeBuilder,
+  commerceInterface: async ({page}, use) => {
+    await use(new CommerceInterface(page));
+  },
+});
+export {expect} from '@playwright/test';